refactor: rename components. Remove js code from flights view
